common time_series helpers
